refactor(ios-app): extract open approval state handling
